% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF.R
\name{KF}
\alias{KF}
\title{Standard Kalman Filter and Smoother}
\usage{
KF(par, yt, mats, delivery_time, dt, smoothing, seasonality)
}
\arguments{
\item{par}{A vector of parameters.}

\item{yt}{The logarihm of futures prices.}

\item{delivery_time}{A vector of date, which is necessary if \code{seasonality} is \code{Constant}.}

\item{dt}{\eqn{\Delta t}. The interval between two consecutive time points.}

\item{smoothing}{Boolean. Indicate if Kalman Smoothing is required.}

\item{seasonality}{\code{Constant} or \code{None}.}

\item{T}{Time to maturity.}
}
\value{
This function returns a list with components:
\item{nll}{The negative log likelihood. }
\item{ll_table}{A vector to store cumulative log-likelihood at each time point.
Used to calculate Sandwich variance. }
\item{table_at_filter}{A matrix gives the filtered values of state variables. }
\item{table_at_prediction}{A matrix gives the predicted values of state variables. }
\item{table_at_smoother}{A matrix gives the smoothed values of state variables.
The algorithm of Kalman Smoother is given by Bierman (1973) and De Jong (1989). }
\item{ft}{Seasonal effect. }
}
\description{
An algorithm for estimating the hidden state variables of a linear dynamic system.
}
\details{
Details.
}
\examples{
###############################################
##### Schwartz and Smith two-factor model #####
###############################################
n_obs <- 100
n_contract <- 10
par <- c(0.5, 0.3, 1, 1.5, 1.3, -0.3, 0.5, 0.3, seq(from = 0.1, to = 0.01, length.out = n_contract))
x0 <- c(0, 1/0.3)
dt <- 1/360 # daily data
n_coe <- 0
func_f <- function(xt, par) state_linear(xt, par, dt) # state equation
dat <- simulate_data(par, x0, n_obs, n_contract, func_f, measurement_linear, n_coe, "Gaussian", 1234)
log_price <- dat$yt # measurement_linear function returns the logarithm of futures price
mats <- dat$mats
est <- KF(par = c(par, x0), yt = log_price, mats = mats, delivery_time = 0, # delivery_time is unnecessary 
          dt = dt, smoothing = FALSE, seasonality = "None")
}
\seealso{
\link{EKF}, \link{UKF} for other filtering methods.
}
