% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_linear.R
\name{measurement_linear}
\alias{measurement_linear}
\title{Linear measurement equation}
\usage{
measurement_linear(x, par, mats)
}
\arguments{
\item{x}{State variables \eqn{x_t}.}

\item{par}{A vector of parameters and model coefficients.}

\item{mats}{Time to maturity.}
}
\value{
This function returns a list with components:
\item{y}{The futures price \eqn{y_t}. }
\item{y_jacobian}{The jacobian of \eqn{y_t}. }
}
\description{
Measurement equation in a linear form. This function can be passed to \link{simulate_data}
as the argument of \code{func_g}.
}
\seealso{
\link{measurement_polynomial}, \link{state_linear} for other forms of state and
measurement equations.
}
