% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_polynomial.R
\name{measurement_polynomial}
\alias{measurement_polynomial}
\title{Polynomial measurement equation}
\usage{
measurement_polynomial(x, par, mats, degree, n_coe)
}
\arguments{
\item{x}{State variables \eqn{x_t}.}

\item{par}{A vector of parameters and model coefficients.}

\item{mats}{Time to maturity.}

\item{degree}{Degree of polynomial.}

\item{n_coe}{The number of model coefficients.}
}
\value{
This function returns a list with components:
\item{y}{The futures price \eqn{y_t}. }
\item{y_jacobian}{The jacobian of \eqn{y_t}. }
}
\description{
Measurement equation in a polynomial form. This function can be passed to \link{simulate_data},
\link{EKF} and \link{UKF} as the argument of \code{func_g}.
}
\seealso{
\link{measurement_linear}, \link{state_linear} for other forms of state and
measurement equations.
}
