% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDSim.R
\docType{package}
\name{PDSim}
\alias{PDSim}
\alias{PDSim-package}
\title{Polynomial diffusion model simulation and estimation}
\description{
Web application for the polynomial diffusion model.
This app generates a futures price data by providing all parameters.
Also, it gives state variables and contracts estimations through
Extended Kalman Filter (EKF) or Unscented Kalman Filter (UKF).
The Schwartz and Smith's two-factor model is also provided for comparison.
}
\author{
Peilun He \href{mailto:peilun.he93@gmail.com}{peilun.he93@gmail.com}

Nino Kordzakhia \href{mailto:nino.kordzakhia@mq.edu.au}{nino.kordzakhia@mq.edu.au}

Gareth Peters \href{mailto:garethpeters@ucsb.edu}{garethpeters@ucsb.edu}

Pavel Shevchenko \href{mailto:pavel.shevchenko@mq.edu.au}{pavel.shevchenko@mq.edu.au}
}
