% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Starts the PDSim app}
\usage{
run_app()
}
\description{
Web application for the polynomial diffusion model.
This app generates a futures price data by providing all parameters.
Also, it gives state variables and contracts estimations through
Extended Kalman Filter (EKF) or Unscented Kalman Filter (UKF).
The Schwartz and Smith's two-factor model is also provided for comparison.
}
\examples{
\dontrun{
library(PDSim)
PDSim::run_app()
}
}
