% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_linear.R
\name{state_linear}
\alias{state_linear}
\title{Linear state equation.}
\usage{
state_linear(x, par, dt)
}
\arguments{
\item{x}{State variables \eqn{x_{t-1}} at previous time point \eqn{t-1}.}

\item{par}{A vector of parameters and model coefficients.}

\item{dt}{\eqn{\Delta t}. The interval between two consecutive time points.}
}
\value{
This function returns a list with components:
\item{y}{The state variables \eqn{x_t} at current time point \eqn{t}. }
\item{y_jacobian}{The jacobian of \eqn{x_t}. }
}
\description{
State equation in a linear form \eqn{x_t = A + B x_{t-1}}. This function can be passed to \link{simulate_data},
\link{EKF} and \link{UKF} as the argument of \code{func_f}.
}
\seealso{
\link{measurement_linear}, \link{measurement_polynomial} for other forms of state and
measurement equations.
}
